+++ app/src/main/java/com/babylon/wallet/android/data/repository/stream/StreamRepository.kt
@@ -64,6 +64,7 @@ class StreamRepositoryImpl @Inject constructor(
         }
     }
 
+    @Suppress("LongMethod")
     private fun buildStreamTransactionRequest(
         accountAddress: String,
         cursor: String?,
