+++ app/src/main/java/com/babylon/wallet/android/utils/DateExtensions.kt
@@ -12,6 +12,7 @@ const val LAST_USED_DATE_FORMAT = "d MMM yyyy"
 const val LAST_USED_DATE_FORMAT_THIS_YEAR = "d MMM"
 const val TIMESTAMP_HOURS_MINUTES = "hh:mm"
 const val TIMESTAMP_ONLY_MONTH = "MMM"
+const val TIMESTAMP_MONTH_AND_YEAR = "MMM YY"
 
 fun LocalDateTime.toISO8601String(): String {
     return DateTimeFormatter.ISO_ZONED_DATE_TIME.format(
@@ -36,7 +37,7 @@ fun String.fromISO8601String(): LocalDateTime? {
     }
 }
 
-fun Instant.ledgerLastUsedDateFormat(): String {
+fun Instant.dayMonthDateFormat(): String {
     val zoneId = ZoneId.systemDefault()
     val currentYear = Instant.now().atZone(zoneId).year
     val instantYear = atZone(zoneId).year
@@ -60,6 +61,12 @@ fun Instant.toDateString(): String {
 }
 
 fun ZonedDateTime.toMonthString(): String {
-    val formatter = DateTimeFormatter.ofPattern(TIMESTAMP_ONLY_MONTH)
+    val currentYear = Instant.now().atZone(zone).year
+    val format = if (currentYear == year) {
+        TIMESTAMP_ONLY_MONTH
+    } else {
+        TIMESTAMP_MONTH_AND_YEAR
+    }
+    val formatter = DateTimeFormatter.ofPattern(format)
     return formatter.format(this)
 }
